pyproject.toml
[tool.poetry]
name = "tg-bot-photohost"
version = "0.1.0"
description = ""
authors = ["Alexander Kolesnikov <alexkolesnikov.git@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.12.0"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
apscheduler = "^3.10.4"


[tool.poetry.group.dev.dependencies]
flake8-isort = "^6.1.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
skip_gitignore = true
src_paths = ["/home/alexander/test_tasks/tg-bot-photohost"]